+++ apps/explorer/lib/explorer/chain/smart_contract/proxy/models/implementation.ex
@@ -94,7 +94,7 @@ defmodule Explorer.Chain.SmartContract.Proxy.Models.Implementation do
   @doc """
   Returns the last implementation updated_at for the given smart-contract address hash
   """
-  @spec get_proxy_implementation_updated_at(Hash.Address.t() | nil, Keyword.t()) :: DateTime.t()
+  @spec get_proxy_implementation_updated_at(Hash.Address.t() | nil, Keyword.t()) :: DateTime.t() | nil
   def get_proxy_implementation_updated_at(proxy_address_hash, options) do
     proxy_address_hash
     |> get_proxy_implementations_query()
@@ -140,9 +140,20 @@ defmodule Explorer.Chain.SmartContract.Proxy.Models.Implementation do
 
     {updated_smart_contract, implementation_address_fetched?} =
       if check_implementation_refetch_necessity(implementation_updated_at) do
-        SmartContract.address_hash_to_smart_contract_with_bytecode_twin(address_hash, options)
+        {smart_contract_with_bytecode_twin, implementation_address_fetched?} =
+          SmartContract.address_hash_to_smart_contract_with_bytecode_twin(address_hash, options)
+
+        if smart_contract_with_bytecode_twin do
+          {smart_contract_with_bytecode_twin, implementation_address_fetched?}
+        else
+          {smart_contract, implementation_address_fetched?}
+        end
       else
-        {smart_contract, false}
+        if implementation_updated_at do
+          {smart_contract, true}
+        else
+          {smart_contract, false}
+        end
       end
 
     get_implementation(